requirements.txt
requests
beautifulsoup4
python-dotenv
lxml
urllib3
argparse